refactor: create separate lib for tools
