+++ packages/app_center/lib/games.dart
@@ -1,3 +1,4 @@
 export 'src/games/games_page.dart';
 export 'src/games/games_page_external_tools.dart';
 export 'src/games/games_page_featured.dart';
+export 'src/games/tool.dart';
