+++ packages/app_center/lib/src/games/games_page.dart
@@ -5,6 +5,7 @@ import 'package:app_center/l10n.dart';
 import 'package:app_center/layout.dart';
 import 'package:app_center/snapd.dart';
 import 'package:app_center/store.dart';
+import 'package:app_center/tool.dart';
 import 'package:app_center/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
