+++ packages/app_center/lib/src/tool/tool.dart
@@ -1,0 +1,9 @@
+class Tool {
+  Tool(this.name, this.summary, this.publisher, this.iconUrl, this.url);
+
+  final String name;
+  final String summary;
+  final String publisher;
+  final String iconUrl;
+  final String url;
+}
