+++ packages/app_center/lib/src/widgets/banner.dart
@@ -3,6 +3,7 @@ import 'package:app_center/layout.dart';
 import 'package:app_center/snapd.dart';
 import 'package:app_center/src/store/store_pages.dart';
 import 'package:app_center/store.dart';
+import 'package:app_center/tool.dart';
 import 'package:app_center/widgets.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
